refactor: location of storage wrappers
